+++ tests/clipboard.py
@@ -342,7 +342,8 @@ some <b>bold</b> text
 		parsetree = page.get_parsetree()
 		Clipboard.set_parsetree(self.notebook, page, parsetree)
 
-		newtree = Clipboard.get_parsetree(self.notebook, page)
+		with tests.LoggingFilter('zim.notebook', 'Could not resolve filename'):
+			newtree = Clipboard.get_parsetree(self.notebook, page)
 		self.assertEqual(newtree.tostring(),
 			'<?xml version=\'1.0\' encoding=\'utf-8\'?>\n'
 			'<zim-tree><p>'
@@ -358,7 +359,8 @@ some <b>bold</b> text
 		parsetree = page.get_parsetree()
 		Clipboard.set_parsetree(self.notebook, page, parsetree)
 
-		newtree = Clipboard.get_parsetree(self.notebook, Path('OtherPage'))
+		with tests.LoggingFilter('zim.notebook', 'Could not resolve filename'):
+			newtree = Clipboard.get_parsetree(self.notebook, Path('OtherPage'))
 		self.assertEqual(newtree.tostring(),
 			'<?xml version=\'1.0\' encoding=\'utf-8\'?>\n'
 			'<zim-tree><p>'
@@ -506,9 +508,9 @@ some <b>bold</b> text
 
 		newpage = self.notebook.get_page(Path('OtherPage'))
 		newpage.attachments_folder.file('foo.png').touch() # Let file exist already
-		newpage.attachments_folder.file('foo001.tex').touch() # Let file exist already
-		newfile_1 = newpage.attachments_folder.file('foo002.png')
-		newfile_2 = newpage.attachments_folder.file('foo002.tex')
+		newpage.attachments_folder.file('foo-001.tex').touch() # Let file exist already
+		newfile_1 = newpage.attachments_folder.file('foo-002.png')
+		newfile_2 = newpage.attachments_folder.file('foo-002.tex')
 		self.assertFalse(newfile_1.exists())
 		self.assertFalse(newfile_2.exists())
 
@@ -520,7 +522,7 @@ some <b>bold</b> text
 			'<?xml version=\'1.0\' encoding=\'utf-8\'?>\n'
 			'<zim-tree><p>'
 			'<object src="%s" type="image+equation" />'
-			'</p></zim-tree>' % os_native_path('./foo002.png')
+			'</p></zim-tree>' % os_native_path('./foo-002.png')
 		) # Sources are copied and number is added due to existing files
 		self.assertTrue(newfile_1.exists())
 		self.assertTrue(newfile_2.exists())