+++ zim/gui/propertiesdialog.py
@@ -19,7 +19,8 @@ notebook_properties = (
 	('icon', 'image', _('Icon')), # T: label for properties dialog
 	('document_root', 'dir', _('Document Root')), # T: label for properties dialog
 	('short_links', 'bool', _('Prefer short names for page links'), False), # T: label for properties dialog
-	('disable_trash', 'bool', _('Do not use system trash for this notebook'), False) # T: label for properties dialog
+	('disable_trash', 'bool', _('Do not use system trash for this notebook'), False), # T: label for properties dialog
+	('paste_image_template', 'string', _('Filename template for pasted images')),
 	# 'shared' property is not shown in properties anymore
 )